% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moranSVGs.R
\name{calculate_moranSVGs}
\alias{calculate_moranSVGs}
\title{calculate_moranSVGs}
\usage{
calculate_moranSVGs(
  seurat_obj,
  sample_name = seurat_obj@project.name,
  output_path = NULL,
  layer = "counts",
  k = 50,
  min_count = 10
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the spatial transcriptomics data for the sample.}

\item{sample_name}{A character string specifying the name of the sample being processed. This will be used in messages and output filenames. Default seurat_obj@project.name.}

\item{output_path}{A character string specifying the directory where the output CSV file will be saved. If \code{NULL}, the results will not be saved to a file. Default is \code{NULL}.}

\item{layer}{A character string indicating the assay layer to use for expression data. Default is \code{"counts"}.}

\item{k}{An integer specifying the number of nearest neighbors to consider. Default is \code{50}.}

\item{min_count}{The minimum number of cells a gene has to be present in (count greater than 0 in a given cell). Default is \code{10}.}
}
\value{
A data frame containing the spatially variable genes and their Moran's I statistics.
}
\description{
Computes spatially variable genes (SVGs) using Moran's I statistic, saves the results to a CSV file if an output path is provided, and returns the results as a data frame.
}
