% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbors_with_weights.R
\name{find_neighbors_with_weights}
\alias{find_neighbors_with_weights}
\title{Calculate Neighbors with Weights}
\usage{
find_neighbors_with_weights(
  seurat_obj,
  method = c("knns", "distance"),
  k = 20,
  dist = 10,
  bin = 8
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{method}{Method for finding neighbors ("knns" or "distance").}

\item{k}{Number of nearest neighbors for "knns".}

\item{dist}{Maximum distance for "distance".}

\item{bin}{Resolution bin size.}
}
\value{
A list containing neighbors and weights.
}
\description{
Calculate Neighbors with Weights
}
