% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots_neighbors.R
\name{hotspot_neighbors}
\alias{hotspot_neighbors}
\title{hotspot_neighbors}
\usage{
hotspot_neighbors(seurat_obj, distance_threshold, metadata_column)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial transcriptomics data.}

\item{distance}{Distance to extend from hotspots to define neighbors}

\item{metada_column}{A metadta column of the Seurat Object which has the "hotspot" classification.}
}
\description{
Identifies the neighbors of hotspots by extending upto a specified radius from the hotspot
}
\examples{
\dontrun{
seurat_obj <- hotspot_neighbors(seurat_obj, distance_threshold = 50, metadata_column = "hotspot_GeneX")
}

}
