% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_line}
\alias{plot_line}
\title{plot function, given two numeric vectors}
\usage{
plot_line(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\description{
plot function, given two numeric vectors
}
